requirements.txt
pandas==2.2.2
numpy==1.26.4
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
faiss-cpu==1.7.4 
streamlit==1.35.0 
Pillow==10.3.0 
requests==2.32.3 
tqdm==4.66.4 
scikit-learn==1.5.0